Share code between admin and participant UIs
