+++ api-participant/build.gradle
@@ -14,7 +14,6 @@ apply from: 'publishing.gradle'
 dependencies {
     implementation 'bio.terra:terra-common-lib'
     implementation project(':core')
-    implementation project(':ui-participant')
     implementation 'org.apache.commons:commons-dbcp2'
     implementation 'org.apache.commons:commons-text:1.10.0'
     implementation 'commons-beanutils:commons-beanutils:1.9.4'
@@ -52,7 +51,7 @@ task copyWebApp(type: Copy) {
 // for now, only jib depends on copyWebApp, so that a npm rebuild/install will not be triggered for
 // development redeploys.  this means to deploy locally with the static assets in place you'll
 // need to run the copyWebApp task yourself
-copyWebApp.dependsOn(':ui-participant:bundle')
+copyWebApp.dependsOn(rootProject.bundleParticipantUI)
 jibDockerBuild.dependsOn('copyWebApp')
 
 test {
