+++ settings.gradle
@@ -1,4 +1,4 @@
 rootProject.name = 'pearl'
-include('core', 'populate', 'ui-admin', 'api-admin', 'api-participant', 'ui-participant')
+include('core', 'populate', 'api-admin', 'api-participant')
 
 gradle.ext.releaseVersion = '0.0.85'
