+++ ui-core/babel.config.json
@@ -1,0 +1,3 @@
+{
+  "presets": [["react-app", { "absoluteRuntime": false }]]
+}
