+++ ui-core/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@juniper/ui-core",
+  "version": "0.0.1",
+  "private": true,
+  "devDependencies": {
+    "@babel/cli": "^7.21.5",
+    "@types/react": "^18.0.18"
+  },
+  "peerDependencies": {
+    "classnames": "^2.3.2",
+    "react": "^18.2.0",
+    "react-router-dom": "^6.4.0"
+  },
+  "main": "./build/index.js",
+  "scripts": {
+    "build": "NODE_ENV=production babel src --copy-files --delete-dir-on-start --extensions .js,.jsx,.ts,.tsx --out-dir build && tsc --declaration --emitDeclarationOnly"
+  }
+}
