+++ ui-core/src/index.ts
@@ -1,0 +1,3 @@
+export { InvestigatorTermsOfUse } from './terms/InvestigatorTermsOfUse'
+export { ParticipantTermsOfUse } from './terms/ParticipantTermsOfUse'
+export { PrivacyPolicy } from './terms/PrivacyPolicy'
