+++ ui-core/tsconfig.json
@@ -1,8 +1,8 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
-    "baseUrl": "src",
-    "outDir": "build",
+    "outDir": "./build",
     "rootDir": "src"
-  }
+  },
+  "include": ["src"]
 }
